chore: Update app.json and package.json
The changes in this commit update the `app.json` and `package.json` files. In `app.json`, the `supportsTablet` property is set to `true` for iOS, and the `bundleIdentifier` property is added with the value "com.abinav-2.hackfest-mobile". For Android, the `adaptiveIcon` property is updated with a new `package` property set to "com.abinav2.hackfest_mobile". Additionally, an `extra` object is added to `app.json` with an `eas` property containing a `projectId` value of "efcde5d6-d874-41d3-afad-fd907e30e85c".

In `package.json`, new dependencies are added for `@react-native-picker/picker`, `@react-navigation/native`, `@react-navigation/native-stack`, `@react-navigation/stack`, `expo-system-ui`, `react-native-linear-gradient`, `react-native-paper`, and `react-native-maps`. The existing `expo` and `expo-status-bar` dependencies are updated to specific versions.

Please remove any meta information and wrap the commit message in a single markdown code block.